Fix Makefile

+++ utils/padding_left.c
@@ -3,12 +3,14 @@
 /*                                                        :::      ::::::::   */
 /*   padding_left.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: besteba <marvin@42.fr>                     +#+  +:+       +#+        */
+/*   By: besteba <besteba@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/09/17 12:02:52 by besteba           #+#    #+#             */
-/*   Updated: 2018/09/17 12:02:54 by besteba          ###   ########.fr       */
+/*   Updated: 2018/09/27 12:28:35 by besteba          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
+
+#include "ft_ls.h"
 
 char		*padding_left(char *str, int width)
 {
@@ -16,10 +18,10 @@ char		*padding_left(char *str, int width)
 	char	*new;
 
 	i = 0;
-	if (width <= ft_strlen(str))
+	if (width <= (int)ft_strlen(str))
 		return (str);
 	new = malloc(sizeof(char) * (width + 1));
-	while (i < width - ft_strlen(str))
+	while (i < width - (int)ft_strlen(str))
 	{
 		new[i] = ' ';
 		i++;
